new flags

+++ Algos/Derangement.cpp
@@ -31,7 +31,7 @@ unsigned long long derangement(const unsigned long long n) {
 
     unsigned long long prev1{1ll};
     unsigned long long prev2{0ll};
-    for (int i = 3; i <= n; i++) {
+    for (unsigned long long i = 3; i <= n; i++) {
 	const auto curr = (i - 1) * (prev1 + prev2);
 	prev2 = prev1;
 	prev1 = curr;
